Let pilight work the JSON double instead of integer

+++ libs/pilight/json.c
@@ -1392,7 +1392,7 @@ bool json_check(const JsonNode *node, char errmsg[256])
 	#undef problem
 }
 
-int json_find_number(JsonNode *object, const char *name, int *out) {
+int json_find_number(JsonNode *object, const char *name, double *out) {
 	JsonNode *node = json_find_member(object, name);
 	if (node && node->tag == JSON_NUMBER) {
 		*out = node->number_;
